Split service point logic into separate client
Closes #34.
