+++ src/Client.php
@@ -2,7 +2,6 @@
 
 namespace JouwWeb\Sendcloud;
 
-use GuzzleHttp\Exception\ConnectException;
 use GuzzleHttp\Exception\RequestException;
 use GuzzleHttp\Exception\TransferException;
 use GuzzleHttp\Utils;
@@ -17,7 +16,6 @@ use JouwWeb\Sendcloud\Model\SenderAddress;
 use JouwWeb\Sendcloud\Model\ShippingMethod;
 use JouwWeb\Sendcloud\Model\User;
 use JouwWeb\Sendcloud\Model\WebhookEvent;
-use JouwWeb\Sendcloud\Model\ServicePoint;
 use Psr\Http\Message\RequestInterface;
 
 /**
@@ -26,7 +24,6 @@ use Psr\Http\Message\RequestInterface;
 class Client
 {
     protected const API_BASE_URL = 'https://panel.sendcloud.sc/api/v2/';
-    public const SERVICE_POINTS_BASE_URL = 'https://servicepoints.sendcloud.sc/api/v2/';
 
     protected \GuzzleHttp\Client $guzzleClient;
 
@@ -67,7 +64,7 @@ class Client
         try {
             return User::fromData(json_decode((string)$this->guzzleClient->get('user')->getBody(), true)['user']);
         } catch (TransferException $exception) {
-            throw $this->parseGuzzleException($exception, 'An error occurred while fetching the Sendcloud user.');
+            throw Utility::parseGuzzleException($exception, 'An error occurred while fetching the Sendcloud user.');
         }
     }
 
@@ -131,7 +128,7 @@ class Client
 
             return $shippingMethods;
         } catch (TransferException $exception) {
-            throw $this->parseGuzzleException(
+            throw Utility::parseGuzzleException(
                 $exception,
                 'An error occurred while fetching shipping methods from the Sendcloud API.'
             );
@@ -197,7 +194,7 @@ class Client
 
             return Parcel::fromData(json_decode((string)$response->getBody(), true)['parcel']);
         } catch (TransferException $exception) {
-            throw $this->parseGuzzleException($exception, 'Could not create parcel in Sendcloud.');
+            throw Utility::parseGuzzleException($exception, 'Could not create parcel in Sendcloud.');
         }
     }
 
@@ -277,7 +274,7 @@ class Client
 
             return $parcels;
         } catch (TransferException $exception) {
-            throw $this->parseGuzzleException($exception, 'Could not create parcel in Sendcloud.');
+            throw Utility::parseGuzzleException($exception, 'Could not create parcel in Sendcloud.');
         }
     }
 
@@ -312,7 +309,7 @@ class Client
 
             return Parcel::fromData(json_decode((string)$response->getBody(), true)['parcel']);
         } catch (TransferException $exception) {
-            throw $this->parseGuzzleException($exception, 'Could not update parcel in Sendcloud.');
+            throw Utility::parseGuzzleException($exception, 'Could not update parcel in Sendcloud.');
         }
     }
 
@@ -348,7 +345,7 @@ class Client
 
             return Parcel::fromData(json_decode((string)$response->getBody(), true)['parcel']);
         } catch (TransferException $exception) {
-            throw $this->parseGuzzleException($exception, 'Could not create parcel with Sendcloud.');
+            throw Utility::parseGuzzleException($exception, 'Could not create parcel with Sendcloud.');
         }
     }
 
@@ -373,7 +370,7 @@ class Client
                 return false;
             }
 
-            throw $this->parseGuzzleException($exception, 'An error occurred while cancelling the parcel.');
+            throw Utility::parseGuzzleException($exception, 'An error occurred while cancelling the parcel.');
         }
     }
 
@@ -405,7 +402,7 @@ class Client
         try {
             return (string)$this->guzzleClient->get($labelUrl)->getBody();
         } catch (TransferException $exception) {
-            throw $this->parseGuzzleException($exception, 'Could not retrieve label.');
+            throw Utility::parseGuzzleException($exception, 'Could not retrieve label.');
         }
     }
 
@@ -440,7 +437,7 @@ class Client
                 ],
             ]);
         } catch (TransferException $exception) {
-            throw $this->parseGuzzleException($exception, 'Could not retrieve label information.');
+            throw Utility::parseGuzzleException($exception, 'Could not retrieve label information.');
         }
 
         $labelData = json_decode((string)$response->getBody(), true);
@@ -452,7 +449,7 @@ class Client
         try {
             return (string)$this->guzzleClient->get($labelUrl)->getBody();
         } catch (TransferException $exception) {
-            throw $this->parseGuzzleException($exception, 'Could not retrieve label PDF data.');
+            throw Utility::parseGuzzleException($exception, 'Could not retrieve label PDF data.');
         }
     }
 
@@ -472,7 +469,7 @@ class Client
                 return SenderAddress::fromData($senderAddressData);
             }, $senderAddressesData);
         } catch (TransferException $exception) {
-            throw $this->parseGuzzleException($exception, 'Could not retrieve sender addresses.');
+            throw Utility::parseGuzzleException($exception, 'Could not retrieve sender addresses.');
         }
     }
 
@@ -487,7 +484,7 @@ class Client
             $response = $this->guzzleClient->get('parcels/' . $this->parseParcelArgument($parcel));
             return Parcel::fromData(json_decode((string)$response->getBody(), true)['parcel']);
         } catch (TransferException $exception) {
-            throw $this->parseGuzzleException($exception, 'Could not retrieve parcel.');
+            throw Utility::parseGuzzleException($exception, 'Could not retrieve parcel.');
         }
     }
 
@@ -520,135 +517,6 @@ class Client
             }
 
             throw $exception;
-        }
-    }
-
-    /**
-     * Summary of searchServicePoints
-     *
-     * @see https://api.sendcloud.dev/docs/sendcloud-public-api/service-points%2Foperations%2Flist-service-points
-     * @param string $country A country ISO 2 code (Example : 'NL')
-     * @param string|null $address Address of the destination address. Can accept postal code instead of the street and the house number. (Example : 'Stadhuisplein 10')
-     * @param string|null $carrier A comma-separated list of carrier codes (stringified) (Example : 'postnl,dpd')
-     * @param string|null $city City of the destination address. (Example : 'Eindhoven')
-     * @param string|null $houseNumber House number of the destination address. (Example : '10')
-     * @param string|null $latitude Used as a reference point to calculate the distance of the service point to the provided location.
-     * @param string|null $longitude Used as a reference point to calculate the distance of the service point to the provided location.
-     * @param string|null $neLatitude Latitude of the northeast corner of the bounding box.
-     * @param string|null $neLongitude Longitude of the northeast corner of the bounding box.
-     * @param string|null $postalCode Postal code of the destination address. Using postal_code will return you service points located around that particular postal code. (Example : '5611 EM')
-     * @param string|null $pudoId DPD-specific query parameter. (<= 7 characters)
-     * @param int|null $radius Radius (in meter) of a bounding circle. Can be used instead of the ne_latitude, ne_longitude, sw_latitude, and sw_longitude parameters to define a bounding box. By default, it’s 100 meters. Minimum value: 100 meters. Maximum value: 50 000 meters.
-     * @param string|null $shopType Filters results by their shop type.
-     * @param string|null $swLatitude Latitude of the southwest corner of the bounding box.
-     * @param string|null $swLongitude Longitude of the southwest corner of the bounding box.
-     * @param float|null $weight Weight (in kg.) of the parcel to be shipped to the service points. Certain carriers impose limits for certain service points that cannot accept parcels above a certain weight limit.
-     * @return ServicePoint[]
-     */
-    public function searchServicePoints(
-        string $country,
-        ?string $address = null,
-        ?string $carrier = null,
-        ?string $city = null,
-        ?string $houseNumber = null,
-        ?string $latitude = null,
-        ?string $longitude = null,
-        ?string $neLatitude = null,
-        ?string $neLongitude = null,
-        ?string $postalCode = null,
-        ?string $pudoId = null,
-        ?int $radius = null,
-        ?string $shopType = null,
-        ?string $swLatitude = null,
-        ?string $swLongitude = null,
-        ?float $weight = null
-    ): array {
-        try {
-            // Construct query array
-            $query = [];
-            $query['country'] = $country;
-
-            if (isset($address)) {
-                $query['address'] = $address;
-            }
-            if (isset($carrier)) {
-                $query['carrier'] = $carrier;
-            }
-            if (isset($city)) {
-                $query['city'] = $city;
-            }
-            if (isset($houseNumber)) {
-                $query['house_number'] = $houseNumber;
-            }
-            if (isset($latitude)) {
-                $query['latitude'] = $latitude;
-            }
-            if (isset($longitude)) {
-                $query['longitude'] = $longitude;
-            }
-            if (isset($neLatitude)) {
-                $query['ne_latitude'] = $neLatitude;
-            }
-            if (isset($neLongitude)) {
-                $query['ne_longitude'] = $neLongitude;
-            }
-            if (isset($postalCode)) {
-                $query['postal_code'] = $postalCode;
-            }
-            if (isset($pudoId)) {
-                $query['pudo_id'] = $pudoId;
-            }
-            if (isset($radius)) {
-                $query['radius'] = $radius;
-            }
-            if (isset($shopType)) {
-                $query['shop_type'] = $shopType;
-            }
-            if (isset($swLatitude)) {
-                $query['sw_latitude'] = $swLatitude;
-            }
-            if (isset($swLongitude)) {
-                $query['sw_longitude'] = $swLongitude;
-            }
-            if (isset($weight)) {
-                $query['weight'] = $weight;
-            }
-
-            // Send request
-            $response = $this->guzzleClient->get('service-points', [
-                'query' => $query,
-            ]);
-
-            // Decode and create ServicePoint objects
-            $json = json_decode((string)$response->getBody(), true);
-
-            $servicePoints = [];
-            foreach ($json as $obj) {
-                $servicePoints[] = ServicePoint::fromData($obj);
-            }
-
-            return $servicePoints;
-        } catch (TransferException $exception) {
-            throw $this->parseGuzzleException($exception, 'Could not retrieve service point.');
-        }
-    }
-
-    /**
-     * Returns service point by ID.
-     *
-     * @see https://api.sendcloud.dev/docs/sendcloud-public-api/service-points%2Foperations%2Fget-a-service-point
-     * @return ServicePoint
-     * @throws SendcloudRequestException
-     */
-    public function getServicePoint(ServicePoint|int $servicePoint): ServicePoint
-    {
-        $servicePointId = $servicePoint instanceof ServicePoint ? $servicePoint->getId() : $servicePoint;
-
-        try {
-            $response = $this->guzzleClient->get('service-points/' . $servicePointId);
-            return ServicePoint::fromData(json_decode((string)$response->getBody(), true));
-        } catch (TransferException $exception) {
-            throw $this->parseGuzzleException($exception, 'Could not retrieve service point.');
         }
     }
 
@@ -811,43 +679,6 @@ class Client
         }
 
         return $parcelData;
-    }
-
-    protected function parseGuzzleException(
-        TransferException $exception,
-        string $defaultMessage
-    ): SendcloudRequestException {
-        $message = $defaultMessage;
-        $code = SendcloudRequestException::CODE_UNKNOWN;
-
-        $responseCode = null;
-        $responseMessage = null;
-        if ($exception instanceof RequestException && $exception->hasResponse()) {
-            $responseData = json_decode((string)$exception->getResponse()->getBody(), true);
-            $responseCode = $responseData['error']['code'] ?? null;
-            $responseMessage = $responseData['error']['message'] ?? null;
-        }
-
-        if ($exception instanceof ConnectException) {
-            $message = 'Could not contact Sendcloud API.';
-            $code = SendcloudRequestException::CODE_CONNECTION_FAILED;
-        }
-
-        // Precondition failed, parse response message to determine code of exception
-        if ($exception->getCode() === 401) {
-            $message = 'Invalid public/secret key combination.';
-            $code = SendcloudRequestException::CODE_UNAUTHORIZED;
-        } elseif ($exception->getCode() === 412) {
-            $message = 'Sendcloud account is not fully configured yet.';
-
-            if (stripos($responseMessage, 'no address data') !== false) {
-                $code = SendcloudRequestException::CODE_NO_ADDRESS_DATA;
-            } elseif (stripos($responseMessage, 'not allowed to announce') !== false) {
-                $code = SendcloudRequestException::CODE_NOT_ALLOWED_TO_ANNOUNCE;
-            }
-        }
-
-        return new SendcloudRequestException($message, $code, $exception, $responseCode, $responseMessage);
     }
 
     // TODO: Remove parseParcelArgument() now we use native unions.
